requirements.txt
configparser
flasgger==0.9.0
Flask==1.0.2
Flask-Cors==3.0.4
Werkzeug==1.0.1
azure.core==1.10.0
flask_swagger_ui
apispec
requests
camelot-py==0.10.1
pyyaml==5.4.1
ghostscript==0.7
urllib3==1.22
mysql-connector==2.2.9